moved localized_urls to its own file
made separate dataset publish function
