+++ hysds/utils.py
@@ -34,10 +34,6 @@ from celery.result import AsyncResult
 from atomicwrites import atomic_write
 from bisect import insort
 
-from billiard import Manager, set_start_method, get_context  # noqa
-from billiard.pool import Pool, cpu_count  # noqa
-
-# import hysds
 from hysds.log_utils import logger, payload_hash_exists
 from hysds.celery import app
 from hysds.es_util import get_grq_es
@@ -187,75 +183,6 @@ def download_file(url, path, cache=False):
                 logger.warning(".osaka.locked.json file found, rolling back...")
                 shutil.rmtree(path + ".osaka.locked.json")
             raise
-
-
-def download_file_async_backoff_handler(b, max_tries=6):
-    """
-    @param b: (Dict) backoff information
-        target: function wrapped by backoff
-        args: (tuple) function arguments
-        kwargs: (Dict) keyword arguments
-            cache: Bool (default False) pull from cache
-            event: (optional) Manager().event()
-        tries: (int) number of tries
-        elapsed: (float) function runtime
-        wait: (float) wait time after error
-        exception: (str) error/traceback raised by the function
-    @param max_tries: maximum number of tries allowed by backoff
-    """
-    tries = b["tries"]
-    kwargs = b["kwargs"]
-    args = b["args"]
-    logger.error("download_file_async failed ({}) {} {}".format(tries, args, kwargs))
-
-    if tries >= max_tries - 1:
-        event = kwargs.get("event", None)
-        if event:
-            event.set()
-        exception = b["exception"]
-        raise exception
-
-
-@backoff.on_exception(
-    backoff.constant,
-    Exception,
-    max_tries=6,
-    interval=5,
-    on_backoff=download_file_async_backoff_handler
-)
-def download_file_async(url, path, cache=False, event=None):
-    """
-    @param url: Str
-    @param path: Str
-    @param cache: Bool (default False) pull from cache
-    @param event: Manager().event() (optional)
-    :return: Dict[Str: any] or None
-        if successful, will return localized data information
-        if None that means a previous task failed and will exit early
-    """
-    if event and event.is_set():
-        logger.warning("Previous localize task failed, skipping %s..." % url)
-        return
-
-    loc_t1 = datetime.utcnow()
-    try:
-        download_file(url, path, cache=cache)
-        loc_t2 = datetime.utcnow()
-        loc_dur = (loc_t2 - loc_t1).total_seconds()
-        path_disk_usage = get_disk_usage(path)
-        return {
-            "url": url,
-            "path": path,
-            "disk_usage": path_disk_usage,
-            "time_start": loc_t1.isoformat() + "Z",
-            "time_end": loc_t2.isoformat() + "Z",
-            "duration": loc_dur,
-            "transfer_rate": path_disk_usage / loc_dur,
-        }
-    except Exception as e:
-        tb = traceback.format_exc()
-        logger.error(tb)
-        raise RuntimeError("Failed to download {}: {}\n{}".format(url, str(e), tb))
 
 
 def find_cache_dir(cache_dir):
@@ -569,58 +496,7 @@ def init_pool_logger():
     logger.addHandler(handler)
 
 
-def localize_urls(job, ctx):
-    """Localize urls for job inputs. Track metrics."""
-    job_dir = job["job_info"]["job_dir"]  # get job info
-
-    async_tasks = []
-    localize_urls_list = job.get("localize_urls", [])
-    num_procs = min(max(cpu_count() - 2, 1), len(localize_urls_list))
-    logger.info("multiprocessing procs used: %d" % num_procs)
-
-    with get_context("spawn").Pool(num_procs, initializer=init_pool_logger) as pool, Manager() as manager:
-        event = manager.Event()
-        for i in localize_urls_list:  # localize urls
-            url = i["url"]
-            path = i.get("local_path", None)
-            cache = i.get("cache", True)
-            if path is None:
-                path = "%s/" % job_dir
-            else:
-                if path.startswith("/"):
-                    pass
-                else:
-                    path = os.path.join(job_dir, path)
-            if os.path.isdir(path) or path.endswith("/"):
-                path = os.path.join(path, os.path.basename(url))
-            dir_path = os.path.dirname(path)
-            makedirs(dir_path)
-
-            async_task = pool.apply_async(download_file_async,
-                                          args=(url, path, ), kwds={"cache": cache, "event": event})
-            async_tasks.append(async_task)
-        pool.close()
-        logger.info("Waiting for dataset localization tasks to complete...")
-        pool.join()
-
-        logger.handlers.clear()  # clearing the queue and removing the handler to prevent broken pipe error
-
-        has_error, err = False, ""
-        for t in async_tasks:
-            if t.successful():
-                result = t.get()
-                if result:
-                    job["job_info"]["metrics"]["inputs_localized"].append(result)
-            else:
-                has_error = True
-                logger.error(t._value)  # noqa
-                err = t._value  # noqa
-        if has_error is True:
-            raise RuntimeError("Failed to download {}".format(err))
-
-    return True  # signal run_job() to continue
-
-
+# TODO: create new function to filter out the .localized datasets
 def find_dataset_json(work_dir):
     """Search for *.dataset.json files."""
